requirements.txt
Django>=4.2
django-taggit>=4.0.0
